Re-org and shuffle
Created data classes for both the ATM and SnowRadar. Moved some of the
external functions into those classes and standardized how time is
handled. UTC time is now leap second corrected relative to the GPS
epoch. Lots to do but we can consider this to be a stable port of most
of the origional code. Will start to build out the geospatial analysis
tools in a dev branch now.