change in utcleap
- utcleap function should only ever expect a single float value.
Allowing it to expect lists or tuples is gonna be a bad time. We can use
list comprehensions or vectorization for quickly swapping gps to utc
timestamps.

- reorganize as_dict() method for OIB/AWI/SnowRadar classes. will
probably have to change it again in the future. New behavious is
barebones dict for SnowRadar objects, and extended plus
subclass-specific dict for OIB/AWI objects.