Elevation compensation
This is the first attempt at elevation compensation. Its only really been tested with the AWI demo file. This replicates mode 3 of the elevation_compensation.m legacy code. Lots to do in terms of establishing how the data is windowed after the compensation but the core is there. The functions will need to be cleaned up and moved over to the snow radar class.