Reader fix
Between the 2016 and 2017 CRESIS datasets in MAT the version changed from v5 to v7. The newer version is close to HDF5 so we needed some additional stuctures to be able to read it. The OIB class should be able to read both types without knowing beforehand which version it is.